# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "list product entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List product entities",
                "operationId": "list-product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Product"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create product",
                "operationId": "create-product",
                "parameters": [
                    {
                        "description": "Product entity",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "get product by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a product entity by ID",
                "operationId": "get-product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update product by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a product entity by ID",
                "operationId": "update-product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product entity",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get product by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a product entity by ID",
                "operationId": "delete-product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Product": {
            "type": "object",
            "properties": {
                "genderID": {
                    "type": "integer"
                },
                "groupOfAge": {
                    "type": "integer"
                },
                "officer": {
                    "type": "integer"
                },
                "productName": {
                    "type": "string"
                },
                "productPaymentOfYear": {
                    "type": "number"
                },
                "productPrice": {
                    "type": "integer"
                },
                "productTime": {
                    "type": "integer"
                }
            }
        },
...
        "ent.Category": {
            "type": "object",
            "properties": {
                "category_name": {
                    "description": "CategoryName holds the value of the \"category_name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CategoryQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.CategoryEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Gender": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GenderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GenderEdges"
                },
                "gender_name": {
                    "description": "GenderName holds the value of the \"gender_name\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.GenderEdges": {
            "type": "object",
            "properties": {
                "genders": {
                    "description": "Genders holds the value of the genders edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Product"
                    }
                }
            }
        },
        "ent.GroupOfAge": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GroupOfAgeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GroupOfAgeEdges"
                },
                "group_of_age_age": {
                    "description": "GroupOfAgeAge holds the value of the \"group_of_age_age\" field.",
                    "type": "string"
                },
                "group_of_age_name": {
                    "description": "GroupOfAgeName holds the value of the \"group_of_age_name\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.GroupOfAgeEdges": {
            "type": "object",
            "properties": {
                "groupageProduct": {
                    "description": "GroupageProduct holds the value of the groupage_product edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Product"
                    }
                }
            }
        },
...
        "ent.Inquiry": {
            "type": "object",
            "properties": {
                "Inquiry_messages": {
                    "description": "InquiryMessages holds the value of the \"Inquiry_messages\" field.",
                    "type": "string"
                },
                "Inquiry_time_messages": {
                    "description": "InquiryTimeMessages holds the value of the \"Inquiry_time_messages\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the InquiryQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.InquiryEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.InquiryEdges": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Category holds the value of the Category edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Category"
                },
                "member": {
                    "description": "Member holds the value of the Member edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Member"
                },
                "officer": {
                    "description": "Officer holds the value of the Officer edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Officer"
                },
                "product": {
                    "description": "Product holds the value of the Product edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Product"
                }
            }
        },
        "ent.Insurance": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the InsuranceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.InsuranceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "insurance_address": {
                    "description": "InsuranceAddress holds the value of the \"insurance_address\" field.",
                    "type": "string"
                },
                "insurance_insurer": {
                    "description": "InsuranceInsurer holds the value of the \"insurance_insurer\" field.",
                    "type": "string"
                },
                "insurance_time_buy": {
                    "description": "InsuranceTimeBuy holds the value of the \"insurance_time_buy\" field.",
                    "type": "string"
                },
                "insurance_time_firstpay": {
                    "description": "InsuranceTimeFirstpay holds the value of the \"insurance_time_firstpay\" field.",
                    "type": "string"
                }
            }
        },
        "ent.InsuranceEdges": {
            "type": "object",
            "properties": {
                "hospital": {
                    "description": "Hospital holds the value of the Hospital edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Hospital"
                },
                "insurancePayment": {
                    "description": "InsurancePayment holds the value of the insurance_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payment"
                    }
                },
                "member": {
                    "description": "Member holds the value of the Member edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Member"
                },
                "officer": {
                    "description": "Officer holds the value of the Officer edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Officer"
                },
                "product": {
                    "description": "Product holds the value of the Product edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Product"
                }
            }
        },
...
        "ent.Officer": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OfficerQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OfficerEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "officer_email": {
                    "description": "OfficerEmail holds the value of the \"officer_email\" field.",
                    "type": "string"
                },
                "officer_name": {
                    "description": "OfficerName holds the value of the \"officer_name\" field.",
                    "type": "string"
                },
                "officer_password": {
                    "description": "OfficerPassword holds the value of the \"officer_password\" field.",
                    "type": "string"
                }
            }
        },
        "ent.OfficerEdges": {
            "type": "object",
            "properties": {
                "officerInquiry": {
                    "description": "OfficerInquiry holds the value of the officer_inquiry edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Inquiry"
                    }
                },
                "officerInsurance": {
                    "description": "OfficerInsurance holds the value of the officer_insurance edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Insurance"
                    }
                },
                "officerPayback": {
                    "description": "OfficerPayback holds the value of the officer_payback edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payback"
                    }
                },
                "officerProduct": {
                    "description": "OfficerProduct holds the value of the officer_product edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Product"
                    }
                },
                "officerRecordinsurance": {
                    "description": "OfficerRecordinsurance holds the value of the officer_recordinsurance edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Recordinsurance"
                    }
                }
            }
        },
        "ent.Payback": {
            "type": "object",
            "properties": {
                "Accountnumber": {
                    "description": "Accountnumber holds the value of the \"Accountnumber\" field.",
                    "type": "string"
                },
                "Transfertime": {
                    "description": "Transfertime holds the value of the \"Transfertime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PaybackQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PaybackEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PaybackEdges": {
            "type": "object",
            "properties": {
                "bank": {
                    "description": "Bank holds the value of the Bank edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bank"
                },
                "member": {
                    "description": "Member holds the value of the Member edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Member"
                },
                "officer": {
                    "description": "Officer holds the value of the Officer edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Officer"
                },
                "product": {
                    "description": "Product holds the value of the Product edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Product"
                }
            }
        },
        "ent.Payment": {
            "type": "object",
            "properties": {
                "account_name": {
                    "description": "AccountName holds the value of the \"account_name\" field.",
                    "type": "string"
                },
                "account_number": {
                    "description": "AccountNumber holds the value of the \"account_number\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PaymentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PaymentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "transfer_time": {
                    "description": "TransferTime holds the value of the \"transfer_time\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Product": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ProductQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ProductEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "product_name": {
                    "description": "ProductName holds the value of the \"product_name\" field.",
                    "type": "string"
                },
                "product_payment_of_year": {
                    "description": "ProductPaymentOfYear holds the value of the \"product_payment_of_year\" field.",
                    "type": "number"
                },
                "product_price": {
                    "description": "ProductPrice holds the value of the \"product_price\" field.",
                    "type": "integer"
                },
                "product_time": {
                    "description": "ProductTime holds the value of the \"product_time\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.ProductEdges": {
            "type": "object",
            "properties": {
                "productGender": {
                    "description": "ProductGender holds the value of the product_gender edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Gender"
                },
                "productGroupage": {
                    "description": "ProductGroupage holds the value of the product_groupage edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GroupOfAge"
                },
                "productInquiry": {
                    "description": "ProductInquiry holds the value of the product_inquiry edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Inquiry"
                    }
                },
                "productInsurance": {
                    "description": "ProductInsurance holds the value of the product_insurance edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Insurance"
                    }
                },
                "productOfficer": {
                    "description": "ProductOfficer holds the value of the product_officer edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Officer"
                },
                "productPayback": {
                    "description": "ProductPayback holds the value of the product_payback edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payback"
                    }
                },
                "productPayment": {
                    "description": "ProductPayment holds the value of the product_payment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payment"
                    }
                },
                "productRecordinsurance": {
                    "description": "ProductRecordinsurance holds the value of the product_recordinsurance edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Recordinsurance"
                    }
                }
            }
        },
        "ent.Recordinsurance": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RecordinsuranceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RecordinsuranceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "recordinsurance_time": {
                    "description": "RecordinsuranceTime holds the value of the \"recordinsurance_time\" field.",
                    "type": "string"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}